ferreteria_refactor/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
python-dotenv
PyQt6
psycopg2-binary
